Backend/requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
fiona==1.9.6
geographiclib==2.0
geopandas==0.14.3
geopy==2.4.1
idna==3.7
iniconfig==2.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyproj==3.6.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
responses==0.25.0
Rtree==1.2.0
shapely==2.0.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
